pmp_agent/requirements.txt
fastapi==0.112.2
uvicorn==0.30.6
pydantic==2.9.2
pyyaml==6.0.2
python-dateutil==2.9.0.post0
jsonschema==4.23.0
